Put components at a separated folder
