+++ frontend/assets/js/components/description_card.js
