+++ frontend/assets/js/components/information_card.js
