+++ frontend/assets/js/components/internal_slider.js
