+++ frontend/package.json
@@ -20,6 +20,7 @@
     "jest": "^20.0.3",
     "material-ui": "^0.18.1",
     "node-sass": "^4.5.3",
+    "react-cellblock": "^3.0.0",
     "react-pure-grid": "^2.1.1",
     "react-test-renderer": "^15.5.4",
     "webpack": "^2.6.0",
